Move org.sonar.core.rule to org.sonar.server.rule
