+++ server/sonar-server/src/main/java/org/sonar/server/issue/IssueQueryService.java
@@ -46,7 +46,7 @@ import org.sonar.api.utils.DateUtils;
 import org.sonar.api.utils.SonarException;
 import org.sonar.api.utils.System2;
 import org.sonar.api.web.UserRole;
-import org.sonar.core.rule.RuleKeyFunctions;
+import org.sonar.server.rule.RuleKeyFunctions;
 import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.component.ComponentDto;
